Allow compilation of SHM on pre-3.11 kernels (#372)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/372

Some users reported build errors on pre-3.11 kernels due to O_TMPFILE not being defined. While I think it's fine to not support those old kernels, we should do so by setting the transport as non-viable at runtime, rather than failing to build. For this case, the fix is easy: we can define the value of that flag ourselves, if it's missing. This will fix the build but will cause a runtime error, which however we'll catch (SHM does try to allocate a shared memory segment during startup and if it fails it marks itself as non-viable).

For clarity I also changed how we do this same thing for MEMFD_CLOEXEC.

Reviewed By: beauby

Differential Revision: D27997830

fbshipit-source-id: 3760aa7dccdcea035116950798859fbaeb833837