Always build CMA regardless of glibc version
Summary:
We used to have a CMake check that verified that we had a glibc wrapper for `process_vm_readv`, which was added in glibc 2.15 (March 2012) which is apparently too recent. However doing that in CMake was a bit ugly (the sources weren't self-contained), and it still had problems: building on a recent glibc (like what we do for PyTorch's binary distributions) and running on an old glibc would still fail. (In general, adding glibc to the mix just adds problems).

In this diff I avoid glibc's wrappers by calling `process_vm_readv` through `syscall`. It means the only requirement is for the kernel versions (both at compile-time and at runtime) to support the syscall. The check for support is made at runtime, and results in non-viability. And the source is self-contained.

Reviewed By: beauby

Differential Revision: D28005559

fbshipit-source-id: 280675cb8f4c9adcefc92bbb574671848a0e92c8