Lift all-to-all restriction in channel selection.
Before support for sender-specified target devices and XDTT, we
assumed that the target device type was the same as the source device
type, which allowed us to pick a channel based solely on the source
device type. This forced us to ensure a given channel connected all
pairs of devices that it supported.
Now that we always gather the target device (either from the sender or
from the receiver), that assumption is lifted, and hence the
restriction can be lifted as well.