Fix dangling references in IBV&GDR when stashing work requests
Summary:
An InfiniBand queue can only hold so many pending requests (we configure its capacity when we create it) and if we were to post more requests to it then it would fail. Which is why we keep track of this and stash excess requests and only submit them when some previous ones finish.

However, when stashing, we were either storing a _reference_ to a _stack-allocated_ `send_wr` or `recv_wr` or, more subtly, we were storing these by value but they contained a pointer to a stack-allocated `sge`. Thus when we tried to submit these values to InfiniBand after stashing them they'd be filled with garbage, leading to a lot of very funny errors.

The solution is to simply put the data we need into another structure, which we pass by value, and build the `send_wr`, `recv_wr` and `sge` on demand just before they're needed.

Reviewed By: beauby

Differential Revision: D32102686

fbshipit-source-id: bb605e84cc0e0593df97e420dbc7e97f702b75cc