Avoid accessing possibly-undefined InfiniBand field in GDR (#416)
Summary:
In case of a Work Completion (WC) that resulted in an error, only some fields are "valid": status, wr_id and qp_num. In particular opcode isn't valid, and yet we were inspecting it to determine whether it was a send or recv request (and increase the corresponding pointer). In case of error, the opcode was neither send nor recv and this triggered an assert. The "solution" is to store the opcode ourselves, separately, in a map indexed by wr_id. We already have such a map, hence we just inster a new field into it.

Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/416

Reviewed By: beauby

Differential Revision: D32720035

Pulled By: lw

fbshipit-source-id: bef94c060f6536bc334ef87f536fb05d7bbc1eb1